Add SampleTest.AspNetCore
