+++ src/SampleTest.AspNetCore/Antiforgery/AntiforgeryOptionsTest.cs
