+++ src/SampleTest.AspNetCore/Antiforgery/AntiforgeryTest.cs
