+++ src/SampleTest.AspNetCore/Authentication/AuthenticationPropertiesTest.cs
