+++ src/SampleTest.AspNetCore/Authentication/HandleRequestResultTest.cs
