+++ src/SampleTest.AspNetCore/Authentication/PropertiesDataFormatTest.cs
