+++ src/SampleTest.AspNetCore/Authentication/PropertiesSerializerTest.cs
