+++ src/SampleTest.AspNetCore/Http.HttpResults/BadRequestOfTValueTest.cs
