+++ src/SampleTest.AspNetCore/Http.HttpResults/HttpContextHelper.cs
