+++ src/SampleTest.AspNetCore/Http.HttpResults/HttpResponseExtensions.cs
