+++ src/SampleTest.AspNetCore/Http.HttpResults/OkOfTValueTest.cs
