+++ src/SampleTest.AspNetCore/Http.HttpResults/ResultsTest.cs
