+++ src/SampleTest.AspNetCore/Http/PathStringTest.cs
