+++ src/SampleTest.AspNetCore/Http/ResultsTest.cs
