+++ src/SampleTest.AspNetCore/Http/TypedResultsTest.cs
