+++ src/SampleTest.AspNetCore/Http/UriHelperTest.cs
