+++ src/SampleTest.AspNetCore/OpenIdConnectOptionsTest.cs
